Cambios en FrontEnd
Solucionado un bug del Drawer:
al renderizar una animación, ocacionalmente intentaba dibujar un frame de más y tiraba error.

Agregadas las Fuentes:
fuentes personalizables, como las que se usan en AO.
por defecto están Verdana y Tahoma (con y sin bordes)
Se pueden editar con CBGF y el "font manager" de la carpeta Util.

Reorganizada la estructura de los assets

+++ core/src/com/gamesharp/jfenix13/graphics/Grh.java
@@ -21,14 +21,14 @@ public class Grh {
     private byte started;
     private short loops;
 
-    public Grh(short index) {
-        this(index, 2);
+    public Grh(int index) {
+        this((short)index, 2);
     }
 
-    public Grh(short index, int started) {
+    public Grh(int index, int started) {
         GrhData grhData = Main.game.assets.getGrhs().getGrhData(index);
         if (index == 0) return;
-        this.index = index;
+        this.index = (short)index;
 
         if (started == 2) {
             if (grhData.getCantFrames() > 1)
